wsgi/openshift/requirements.txt
Django==1.6.5
GDAL==1.11.0
Pillow==2.4.0
Pygments==1.6
South==0.8.4
argparse==1.2.1
diff-match-patch==20121119
dj-database-url==0.3.0
dj-static==0.0.5
django-appconf==0.6
django-filebrowser==3.5.3
django-geoposition==0.2.0
django-grappelli==2.5.3
django-leaflet==0.13.2
django-toolbelt==0.0.1
djangorestframework==2.3.14
docutils==0.11
geopy==0.99
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
requests==2.3.0
six==1.6.1
static==1.0.2
tablib==0.10.0
vboxapi==1.0
virtualenv==1.11.4
wsgiref==0.1.2
